Add files via upload
Project statement: 
AAL, established in 2000, is a well-known brand in Australia, particularly 
recognized for its clothing business. It has opened branches in various states, 
metropolises, and tier-1 and tier-2 cities across the country. 
The brand caters to all age groups, from kids to the elderly. 
Currently experiencing a surge in business, AAL is actively pursuing expansion 
opportunities. To facilitate informed investment decisions, the CEO has assigned 
the responsibility to the head of AAL’s sales and marketing (S&M) department. 
The specific tasks include: 
1) Identify the states that are generating the highest revenues. 
2) Develop sales programs for states with lower revenues. The head of sales 
and marketing has requested your assistance with this task. 
Analyze the sales data of the company for the fourth quarter in Australia, 
examining it on a state-by-state basis. Provide insights to assist the company in 
making data-driven decisions for the upcoming year.